map distance

+++ database/migrations/2022_09_18_031355_create_bookings_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->foreign('driver_id')->references('id')->on('users');
             $table->date('trip_start_date')->nullable();
             $table->date('trip_end_date')->nullable();
-            $table->integer('t_total_distance')->default(0);
+            $table->double('t_total_distance')->default(0.0);
             $table->timestamps();
         });
     }
